refactor(accounts): move out of integrations/ + address review nits (#266)
- Move rudderstack/integrations/accounts → rudderstack/accounts (accounts
  aren't integrations); repoint the registration import + Makefile/CI paths.
- account_bigquery_test: omit optional 'location' from the Create step
  (exercised in Update).
- registration test: also assert field Types and that Properties (API
  mappings) are wired.
- resource_account: drop unused cm arg from resourceAccountDelete.


🔒 Scanned for secrets using gitleaks 8.28.0
