+++ rudderstack/accounts/account_bigquery_registration_test.go
@@ -3,10 +3,11 @@ package accounts_test
 import (
 	"testing"
 
+	sdkschema "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/stretchr/testify/require"
 
 	"github.com/rudderlabs/terraform-provider-rudderstack/rudderstack/configs"
-	_ "github.com/rudderlabs/terraform-provider-rudderstack/rudderstack/integrations/accounts"
+	_ "github.com/rudderlabs/terraform-provider-rudderstack/rudderstack/accounts"
 )
 
 func TestBigQueryAccountRegistration(t *testing.T) {
@@ -35,4 +36,12 @@ func TestBigQueryAccountRegistration(t *testing.T) {
 	require.True(t, ok, "ConfigSchema must contain 'credentials'")
 	require.True(t, credentials.Required, "'credentials' must be Required")
 	require.True(t, credentials.Sensitive, "'credentials' must be Sensitive")
+
+	// Field types
+	require.Equal(t, sdkschema.TypeString, project.Type, "'project' must be a string")
+	require.Equal(t, sdkschema.TypeString, location.Type, "'location' must be a string")
+	require.Equal(t, sdkschema.TypeString, credentials.Type, "'credentials' must be a string")
+
+	// API field mappings (Properties) must be wired for project, location and credentials.
+	require.Len(t, cm.Properties, 3, "expected project, location and credentials property mappings")
 }
