+++ rudderstack/integrations/integrations.go
@@ -1,7 +1,7 @@
 package integrations
 
 import (
-	_ "github.com/rudderlabs/terraform-provider-rudderstack/rudderstack/integrations/accounts"
+	_ "github.com/rudderlabs/terraform-provider-rudderstack/rudderstack/accounts"
 	_ "github.com/rudderlabs/terraform-provider-rudderstack/rudderstack/integrations/destinations"
 	_ "github.com/rudderlabs/terraform-provider-rudderstack/rudderstack/integrations/sources"
 )
